Created edit mode to change existing pages

+++ pages/views.py
@@ -19,6 +19,7 @@ def page(request, page_id):
     template = loader.get_template('pages/page_display.html')
     selected_page = Pages.objects.get(page_id=page_id)
     context = {
+        'page_id': page_id,
         'page_name': selected_page.page_name,
         'page_data': json.loads(selected_page.file.read())
     }
@@ -26,20 +27,26 @@ def page(request, page_id):
 
 
 def add_page(request):
+    return edit_page(request, None)
+
+
+def edit_page(request, page_id):
     template = loader.get_template('pages/page_add.html')
-    context = {}
+    context = {
+        'page_id': json.dumps(page_id)
+    }
     return HttpResponse(template.render(context, request))
 
 
 def get_page_count_ajax(request):
     count = Pages.objects.count()
-    return JsonResponse({"count": count})
+    return JsonResponse({'count': count})
 
 
 def get_pages_ajax(request):
     pages = list(Pages.objects.values_list('page_id', 'page_name', flat=False).order_by('page_id', 'page_name')
-                 [int(request.POST.get("start", 0)):int(request.POST.get("stop", 10))])
-    return JsonResponse({"pages": pages})
+                 [int(request.POST.get('start', 0)):int(request.POST.get('stop', 10))])
+    return JsonResponse({'pages': pages})
 
 
 def save_ajax(request):
@@ -48,14 +55,38 @@ def save_ajax(request):
     Respond with either success or failure message
     """
     if request.method == 'POST':
-        page_data = json.loads(request.POST.get('page data', None).replace("'", '"'))
-        new_page = Pages(page_name=page_data['name'])
-        new_page.save()  # Save to auto generate id for use in file name
-        filename = str(new_page.page_id) + '.json'
-        with open(filename, 'w+') as f:
-            f.write(str(page_data['records']).replace("'", '"'))
-            new_page.file = File(f)
-            new_page.save()
-        os.remove(filename)
-        return JsonResponse({"page_id": str(new_page.page_id)})
-    return JsonResponse({"responseText": "Invalid method " + request.method}, status=500)
+        page_data = json.loads(request.POST.get('page_data', None))
+        page_id = request.POST.get('page_id', None)
+        if page_id:
+            overwrite_page = Pages.objects.get(page_id=page_id)
+            with open(overwrite_page.file.path, "w+") as f:
+                f.write(str(page_data['records']))
+            return JsonResponse({'page_id': str(page_id), 'overwrite': True})
+        else:
+            new_page = Pages(page_name=page_data['name'])
+            new_page.save()  # Save to auto generate id for use in file name
+            filename = str(new_page.page_id) + '.json'
+            with open(filename, 'w+') as f:
+                f.write(json.dumps(page_data['records']))
+                new_page.file = File(f)
+                new_page.save()
+            os.remove(filename)
+            return JsonResponse({'page_id': str(new_page.page_id), 'overwrite': False})
+    return JsonResponse({'responseText': 'Invalid method ' + request.method}, status=500)
+
+
+def get_page_data_ajax(request):
+    if request.method == 'POST':
+        page_id = request.POST.get('page_id', 'null')
+        if page_id == 'null':
+            return JsonResponse({
+                'page_name': '',
+                'page_data': ''
+            })
+        selected_page = Pages.objects.get(page_id=page_id)
+        if selected_page:
+            return JsonResponse({
+                'page_name': selected_page.page_name,
+                'page_data': json.loads(selected_page.file.read())
+            })
+    return JsonResponse({"responseText": 'Invalid method ' + request.method}, status=500)
